dream/dream-business-analysis-ai/requirements.txt
# Core LangChain and LLM dependencies
langchain==0.2.16
langchain-community==0.2.16
langchain-text-splitters==0.2.4
langchain-ollama==0.1.3
langchain-huggingface==0.0.3

# Vector Database
chromadb==0.4.24
langchain-chroma==0.1.4
faiss-cpu==1.8.0

# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2

# Embeddings
sentence-transformers==3.0.1
transformers==4.42.4

# Utilities
PyYAML==6.0.1
requests==2.32.3
Jinja2==3.1.4
aiofiles==23.2.1
pandas==2.2.2
numpy==1.26.4
numpy-financial==1.0.0

# Visualization
matplotlib==3.9.1
plotly==5.22.0

# PDF Generation
fpdf2==2.7.9

# Additional utilities
python-multipart==0.0.9
python-dotenv==1.0.1
streamlit==1.39.0
streamlit-option-menu==0.4.0
plotly==5.22.0